Refactor public score view
